+++ src/main/java/implementations/Report2RowMapper.java
@@ -3,7 +3,6 @@ package implementations;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 
-import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -15,7 +14,7 @@ public class Report2RowMapper implements RowMapper<Report2> {
     @Override
     public Report2 mapRow(ResultSet row, int rowNum) throws SQLException{
         Report2 rep2 = new Report2(jdbcTemplate);
-        rep2.setId(BigInteger.valueOf(row.getInt("ID")));
+        rep2.setId(row.getInt("ID"));
         rep2.setName(row.getString("NAME"));
         rep2.setRequired(row.getBoolean("REQUIRED"));
         rep2.setSum(row.getDouble("SUMMA"));
