+++ src/main/java/interfaces/Budget.java
@@ -1,11 +1,10 @@
 package interfaces;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.Date;
 
 public interface Budget extends Entity {
-    BigInteger getBudgetId();
+    Integer getBudgetId();
     String getOperationType();
     Integer getBudgetTypeId();
     String getDescription();
