+++ src/main/java/interfaces/PlanBudget.java
@@ -1,11 +1,10 @@
 package interfaces;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.Date;
 
 public interface PlanBudget extends Entity {
-    BigInteger getPlanBudgetId();
+    Integer getPlanBudgetId();
     String getOperationType();
     Integer getBudgetTypeId();
     String getDescription();
